Refactor Docker Compose configuration and update i18n imports
